+++ core/src/main/towers/Tank.java
@@ -19,7 +19,7 @@ public class Tank extends Tower {
     private static final float DAMAGE = 100;
     private static final float SPEED = 35;
 
-    public Tank(int x, int y, float radius, float cooldown, int cost, float damage){
+    public Tank(int x, int y, float radius, float cooldown, float cost, float damage){
     	super(x, y, radius, NAME, cost, cooldown, SIZE, damage, PARTY);
     }
